app/requirements.txt
# Core dependencies - Windows compatible versions
psycopg[binary]>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
SQLAlchemy>=2.0.0

# Configuration and environment
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0

# Data validation (lightweight)
pydantic>=2.0.0

# Logging
structlog>=23.0.0
colorama>=0.4.6

# Development essentials
pytest>=7.0.0
black>=23.0.0

schedule==1.2.0